chore: fix mod structure
This patch fixes the module structure so it uses actually
meaningful names for modules. The old structure used a
directory-per-module structure, which makes every buffer in
your editor show up as "mod.rs", which is terrible. This makes
the codebase much more navigable.
