+++ src/descriptor.rs
@@ -1,5 +1,9 @@
-use std::error::Error;
-use std::fmt;
+use std::{
+    error::Error,
+    ffi::CStr,
+    fmt,
+    os::fd::{FromRawFd, OwnedFd},
+};
 
 /// The enum `DescriptorError` defines the possible errors
 /// from constructor Descriptor.
@@ -31,5 +35,21 @@ impl Error for DescriptorError {
     /// any.
     fn cause(&self) -> Option<&dyn Error> {
         None
+    }
+}
+
+/// Open the given file as a file descriptor.
+pub fn open(
+    path: &CStr,
+    flag: libc::c_int,
+    mode: Option<libc::c_int>,
+) -> Result<OwnedFd, DescriptorError> {
+    // Safety: we've just ensured that path is non-null and the
+    // other params are valid by construction.
+    unsafe {
+        match libc::open(path.as_ptr().cast(), flag, mode.unwrap_or_default()) {
+            -1 => Err(DescriptorError::OpenFail),
+            fd => Ok(OwnedFd::from_raw_fd(fd)),
+        }
     }
 }
