+++ src/fork/pty/master.rs
@@ -1,20 +1,57 @@
-mod err;
-
 #[cfg(target_os = "macos")]
 mod ptsname_r_macos;
 
-use crate::descriptor;
-
-pub use self::err::{MasterError, Result};
+use crate::descriptor::{self, DescriptorError};
 use std::{
+    error::Error,
     ffi::CStr,
-    io,
+    fmt, io,
     os::{
         fd::{AsFd, AsRawFd, BorrowedFd, OwnedFd},
         unix::io::RawFd,
     },
     sync::Arc,
 };
+
+/// The alias `Result` learns `MasterError` possibility.
+pub type Result<T> = ::std::result::Result<T, MasterError>;
+
+/// The enum `MasterError` defines the possible errors from constructor Master.
+#[derive(Clone, Copy, Debug)]
+pub enum MasterError {
+    BadDescriptor(DescriptorError),
+    GrantptError,
+    UnlockptError,
+    PtsnameError,
+}
+
+impl fmt::Display for MasterError {
+    /// The function `fmt` formats the value using the given formatter.
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", ::errno::errno())
+    }
+}
+
+impl Error for MasterError {
+    /// The function `description` returns a short description of the error.
+    fn description(&self) -> &str {
+        match *self {
+            MasterError::BadDescriptor(_) => "the descriptor as occured an error",
+            MasterError::GrantptError => "the `grantpt` has a error, errnois set appropriately.",
+            MasterError::UnlockptError => "the `grantpt` has a error, errnois set appropriately.",
+            MasterError::PtsnameError => "the `ptsname` has a error",
+        }
+    }
+
+    /// The function `cause` returns the lower-level cause of this error, if
+    /// any.
+    fn cause(&self) -> Option<&dyn Error> {
+        match *self {
+            MasterError::BadDescriptor(ref err) => Some(err),
+            _ => None,
+        }
+    }
+}
 
 #[derive(Debug, Clone)]
 pub struct Master {
@@ -76,7 +113,8 @@ impl Master {
             let result = libc::ptsname_r(self.raw_fd(), data as *mut libc::c_char, buf.len());
 
             #[cfg(target_os = "macos")]
-            let result = ptsname_r_macos::ptsname_r(fd, data as *mut libc::c_char, buf.len());
+            let result =
+                ptsname_r_macos::ptsname_r(self.raw_fd(), data as *mut libc::c_char, buf.len());
 
             match result {
                 0 => Ok(()),
